refactor(triggers): 将 `chat-send` 也移至 `src/triggers/` 统一管理
